converted consistency check into a unit test
