+++ IPA KeyboardTests/IPA_KeyboardTests.swift
@@ -2,35 +2,32 @@
 //  IPA_KeyboardTests.swift
 //  IPA KeyboardTests
 //
-//  Created by Vincent C. on 2018/9/6.
-//  Copyright © 2018 Vince14Genius. All rights reserved.
+//  Created by Vincent C. on 7/30/23.
+//  Copyright © 2023 Vince14Genius. All rights reserved.
 //
 
 import XCTest
 @testable import IPA_Keyboard
 
-class IPA_KeyboardTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
+final class IPA_KeyboardTests: XCTestCase {
+
+    override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
-    
-    override func tearDown() {
+
+    override func tearDownWithError() throws {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
     }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
+
+    func testKeySetConsistency() throws {
+        KeySetConsistentAcrossSizesTest.check(layout: IPASymbols.self)
     }
-    
-    func testPerformanceExample() {
+
+    func testPerformanceExample() throws {
         // This is an example of a performance test case.
-        self.measure {
+        measure {
             // Put the code you want to measure the time of here.
         }
     }
-    
+
 }
